Fixed version_message
Inclusion of a version_message seems to function correctly. Overall,
it feels wrong or like a hack. There must be a cleaner way to allow
a user to attached a message with a change at the time of its
occurrence instead of committing the change (and History table creation)
then querying it and adding a message as a second step.

Also added a test for the message.