Added test for autogenerated migration
Runs a first migration autogeneration script to see that
the versioned table and associated history table are set
up correctly to be automigrated. Still need to finish
setup/takedown of test so no files are left behind, but
wanted to back it up now that it actually creates a migration.

Specific uses for this test will probably require adapations,
as automigrations aren't intended to be fully auto and do require
intervention by user for certain cases.